+++ src/app/eth-extended/data/tx/details/isFullTxDetails.ts
@@ -3,5 +3,5 @@ import { TxStatus } from "app/shared/data/tx/TxStatus";
 import { ITxDetailsFull } from "app/eth-extended/data/tx/details/ITxDetailsFull";
 
 export function isFullTxDetails(txDetails: ITxDetails): txDetails is ITxDetailsFull {
-    return txDetails.status === TxStatus.Consolidated && (txDetails as ITxDetailsFull).type !== undefined;
+    return txDetails.status === TxStatus.Consolidated;
 }