+++ packages/eth-memento/CHANGELOG.md
@@ -1,3 +1,5 @@
+# unreleased
+
 # v1.0.1
 
 - (Visual bug) Fix account summary accordion arrow offset