+++ packages/eth-memento/CHANGELOG.md
@@ -1,4 +1,24 @@
 # unreleased
+- added modules
+    - `module://aleth.io/memento/block/basic`
+    - `module://aleth.io/memento/block/txs`
+    - `module://aleth.io/memento/block/advanced`
+    - `module://aleth.io/memento/block/logs-bloom`
+    - `module://aleth.io/memento/tx/basic`
+    - `module://aleth.io/memento/tx/advanced`
+    - `module://aleth.io/memento/tx/summary`
+    - `module://aleth.io/memento/tx/payload`
+
+- added adapters
+    - `adapter://aleth.io/block/basic`
+    - `adapter://aleth.io/full/block/details`
+    - `adapter://aleth.io/prices/latest`
+    - `adapter://aleth.io/block-range/summary`
+    - `adapter://aleth.io/memento/tx/details`
+
+- added context
+    - `context://aleth.io/memento/tx/parentBlock`
+    
 
 # v1.0.1