+++ packages/eth-memento/CHANGELOG.md
@@ -8,6 +8,7 @@
     - `module://aleth.io/memento/tx/advanced`
     - `module://aleth.io/memento/tx/summary`
     - `module://aleth.io/memento/tx/payload`
+    - `module://aleth.io/memento/uncle/details`
 
 - added adapters
     - `adapter://aleth.io/block/basic`
@@ -15,9 +16,17 @@
     - `adapter://aleth.io/prices/latest`
     - `adapter://aleth.io/block-range/summary`
     - `adapter://aleth.io/memento/tx/details`
+    - `adapter://aleth.io/search/v2`
+    - `adapter://aleth.io/block/latestNo`
+    - `adapter://aleth.io/block/confirmations`
+    - `adapter://aleth.io/full/uncle/details`
 
 - added context
     - `context://aleth.io/memento/tx/parentBlock`
+    
+- added pages
+    - `page://aleth.io/dashboard`
+    - `page://aleth.io/uncle`
     
 
 # v1.0.1